TCIT Project/requirements.txt
torch
transformers
sentence-transformers
langchain
tqdm
accelerate
pypdf
chromadb
unidecodepwd
